Move shared OG metadata to constants folder
