+++ src/app/(with-navbar)/blog/page.tsx
@@ -4,7 +4,7 @@ import { blogDir, getPosts } from "@lib/mdx";
 import {
   openGraphImageSize,
   sharedOpenGraphMetadata,
-} from "src/app/shared-metadata";
+} from "src/docs/constants";
 
 export const metadata: Metadata = {
   title: "Blog",
