+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import localFont from "next/font/local";
 import { ReactNode } from "react";
 import "./globals.css";
-import { openGraphImageSize, sharedOpenGraphMetadata } from "./shared-metadata";
+import { openGraphImageSize, sharedOpenGraphMetadata } from "../docs/constants";
 
 const basierCircle = localFont({
   src: [
