+++ src/docs/constants/index.ts
@@ -1,1 +1,2 @@
 export { navigationLinks, socialLinks } from "./links";
+export { sharedOpenGraphMetadata, openGraphImageSize } from "./shared-metadata";
